Refactor select numeric columns & debug via logging
Replace if >50% of rows are numeric with regex
that claims column is numeric if any row is a
valid number according to a regex. This will break
if a string column contains a row that is just a number
such as "12" entered in error... >50% of rows failed with
existing regex as empty rows were not counted as numbers
so a mostly empty numeric row would be "not numeric

debug etl flow parameter changes logging level to
debug which provides more feedback on names of numeric
and non-numeric columns for one. Previously debug
activated a ipdb in a function of choice which required debug
to be passed between functions as an argument...
Easier to just drop an ipdb in following strange behaviour
recording by debug logger