Refactor load to excel & transform excels into one task
Previously had a task for each sheet for transform
and load.  Replacing transform with a mapped task
is cleaner & a single to excel task is 10x faster!
It also enables use of Dask as now all tasks can be
run in parallel (loading separate tasks to excel was
the barrier to this before)